refactoring (rarely) evil regex, caching regex to this.data
a regex like:
  /^(\p{Emoji}|[\w\s\d.-/:?!])+$/ui

with a message like:
  BloodTrail 21:37 BloodTrail 21:37 BloodTrail 21:37 BloodTrail 21:37 BloodTrail 21:37 BloodTrail 21:37 BloodTrail 21:37 BloodTrail 21:37 BloodTrail 21:37 BloodTrail 21:37 BloodTrail 21:37 BloodTrail 21:37 BloodTrail 21:37 '

becomes evil when a character outside of its range is at the end (rather, at a high index) of the message. back-trakcing occurs and an exponential calc limit is reached quickly.

regex has been replaced by one that simply checks one character range instead of using the "|" operator to join two groups.